Mobile Application/src/test/TaskTest.java
/************************************************************************************
 *                             Mobile Application
 * 
 * Description: A mobile application for organizing client-related contacts, tasks, 
 * and appointments.
 * Alyssa Fasciano
 * Southern New Hampshire University
 * CS-499: Computer Science Capstone
 * March/April 2023
 ************************************************************************************/
package test;

import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import task.Task;

// JUnit testing for Task class

public class TaskTest {
	
	@Test
	void testTask() {
		Task task = new Task("TaskID", "John", "Description Test");
		assertTrue(task.getTaskID().equals("TaskID"));
		assertTrue(task.getName().equals("John"));
		assertTrue(task.getDescription().equals("Description Test"));
	}
	
	@Test
	void testTaskIDTooLong() {
		Assertions.assertThrows(IllegalArgumentException.class, () -> {
			new Task("TaskIDTaskID", "John", "Description Test");
		});
	}
	
	@Test
	void testTaskIDIsNull() {
		Assertions.assertThrows(IllegalArgumentException.class, () -> {
			new Task(null, "John", "Description Test");
		});
	}
	
	@Test
	void testNameTooLong() {
		Assertions.assertThrows(IllegalArgumentException.class, () -> {
			new Task("TaskID", "JohnJohnJohnJohnJohnJohn", "Description Test");
		});
	}
	
	@Test
	void testNameIsNull() {
		Assertions.assertThrows(IllegalArgumentException.class, () -> {
			new Task("TaskID", null, "Description Test");
		});
	}
	
	@Test
	void testDescriptionTooLong() {
		Assertions.assertThrows(IllegalArgumentException.class, () -> {
			new Task("TaskID", "John", "DescriptionDescriptionDescriptionDescriptionDescription");
		});
	}
	
	@Test
	void testDescriptionIsNull() {
		Assertions.assertThrows(IllegalArgumentException.class, () -> {
			new Task("TaskID", "John", null);
		});
	}

}
